Determine debiasing only using valid samples
This is an attempt to modify Lee's notebook (based on Ross's GAMA
branch) to only use galaxies meeting the criteria to be meaningful to
consider for a given question (pq > 0.5 & Nq > 5). Only such galaxies
are used to compute the correction for each question, and only these
galaxies have their vote fractions debiased. Vote fractions for
galaxies that do not meet the criteria are set to a value of -999.