Include debiased fractions for all galaxies.
Previously we were masking debiased fractions for galaxies that didn't meet the 'prior sample' criteria. However, it seems this can cause difficulties when trying to compute morphology statistics for a full population. Some of the previously masked galaxies still seem to have reasonable vote fractions. We therefore now provide debiased fractions for everything with at least one vote for a question. This is the same approach taken in producing Ross's GZ2 catalogue. In order to allow users to select 'safe' galaxies, we provide a column containing the 'prior sample' mask.

The primary diagnostic plot is unaffected, since we now apply the 'prior sample' mask to the objects used in this plot, consistent with Ross's GZ2 version of the plot.